refactor(gui): extract the serial-debug session into serial_debug.rs
Fourth cut at lib.rs. Moves the twelve serial-debug commands, DebugState, the
chunk bridge that coalesces device output before it crosses the IPC boundary,
the filter payload types, and the device-reset entry points.

The bridge is the reason these belong together: device output arrives
line-by-line but reaches the webview batched by SERIAL_DEBUG_CHUNK_FLUSH_*, so a
chatty device cannot flood the channel. That invariant was previously split
across two regions 1000 lines apart.

Also drops the now-unused tyutool_core import block from lib.rs, which had
carried sixteen serial-debug types.

lib.rs 2561 -> 1873. 45 commands registered, 363 workspace tests pass, clippy
clean, fmt clean.
